requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
cobble==0.1.3
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.9
docx2txt==0.7
idna==2.8
lxml==4.3.0
more-itertools==5.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nltk==3.4
pdfminer.six==20181108
plac==0.9.6
pluggy==0.8.1
preshed==2.0.1
py==1.7.0
pycryptodome==3.7.3
pytest==4.1.1
python-docx==0.8.10
regex==2018.1.10
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
sortedcontainers==2.1.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.29.1
ujson==1.35
urllib3==1.24.1
wrapt==1.10.11